requirements.txt
annotated-types>=0.5.0
anyio>=3.7.1
apprise>=1.6.0
black>=23.3.0
certifi>=2023.11.17
cffi>=1.15.1
charset-normalizer>=2.0.12
click>=8.1.7
coloredlogs>=15.0.1
distro>=1.8.0
exceptiongroup>=1.2.0
flatbuffers>=23.5.26
h11>=0.14.0
httpcore>=0.17.3
httpx>=0.24.1
humanfriendly>=10.0
idna>=3.6
importlib-metadata>=6.7.0
Markdown>=3.5.1
mpmath==1.3.0
mypy-extensions==1.0.0
numpy>=1.21.6,<2.0.0
oauthlib==3.2.2
onnxruntime>=1.16.1
openai>=1.3.7
packaging>=23.2
pathspec>=0.11.2
Pillow>=10.1.0
platformdirs>=4.0.0
playsound==1.2.2
protobuf>=4.24.4
pvporcupine>=3.0.1
pvrecorder>=1.2.1
PyAudio>=0.2.11
pycparser>=2.21
pydantic>=2.5.2
pydantic-core>=2.14.5
python-vlc>=3.0.20123
PyYAML>=6.0.1
requests>=2.26.0
requests-oauthlib>=1.3.1
ruamel.yaml>=0.18.5
ruamel.yaml.clib>=0.2.8
sniffio>=1.3.0
sounddevice>=0.4.6
SpeechRecognition>=3.9.0
sympy>=1.10.1
tomli>=2.0.1
tqdm>=4.66.1
typing-extensions>=4.7.1
urllib3>=1.26.18,<2.0.0
webuiapi>=0.1.0
zipp>=3.15.0